Add PoET BlockPublisher.finalize_block interface shim
The consensus engine API is slightly different from the old consensus
interface. Instead of accepting a block and executing side effects on
it to finalize, the new finalize interface accepts a block summary (a
hash of its contents) and returns the consensus payload, which will
get inserted into a block elsewhere.

Here, the new interface is surreptitiously inserted into the old one.
if the input to the function is bytes, then it is assumed that the new
interface is being used; otherwise, the old interface is used. This
isn't particularly elegant, but it is unobtrusive.

Signed-off-by: Nick Drozd <drozd@bitwise.io>